backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0
pydantic>=2.7
pydantic-settings>=2.2
anthropic>=0.25.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.128.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
httpx>=0.27.0
html2text>=2024.2.26
